Update CREATE TABLE perfmon.sql
Changed tx and rx bytes data types to 'bigint' to handle large integers which are out of range for normal int types in python and sql

+++ CREATE TABLE perfmon.sql
@@ -10,8 +10,8 @@ CREATE TABLE perfmon (
 	cpu_max_temp varchar(15) NULL,
 	disk_usage varchar(10) NULL,
 	ram_usage varchar(10) NULL,
-	tx_bytes int4 NULL,
-	rx_bytes int4 NULL,
+	tx_bytes bigint NULL,
+	rx_bytes bigint NULL,
 	cpu_meas_per_min int4 NULL,
 	CONSTRAINT perfmon_pkey PRIMARY KEY (id)
 );
